Updated arg1 variable to month-date-year format
Previous code used Hour/Minute format.  This would result in log files being overwritten if the script completed at the same time each time it was run.  Now the date stamp (MM-DD-YYYY) will be part of the filename.  This should allow for a new log file each time the script is run, assuming it is run no more frequently than once a day.